+++ RBLX Plugin/src/StarterGui/Argon/Main/HttpHandler.lua
@@ -1,4 +1,5 @@
 local HttpService = game:GetService('HttpService')
+local MarketPlaceService = game:GetService('MarketplaceService')
 
 local FileHandler = require(script.Parent.FileHandler)
 local TwoWaySync = require(script.Parent.TwoWaySync)
@@ -10,6 +11,7 @@ local SYNC_INTERVAL = 0.2
 
 local thread = nil
 local func = nil
+local title = nil
 
 local httpHandler = {}
 httpHandler.file = nil
@@ -75,10 +77,11 @@ end
 
 function httpHandler.connect(fail)
     local url = URL:format(Config.host, Config.port)
-    local header = {action = 'init'}
+    local initHeader = {action = 'init'}
+    local titleHeader = {action = 'setTitle'}
 
     local success, response = pcall(function()
-        if HttpService:GetAsync(url, false, header) == 'true' then
+        if HttpService:GetAsync(url, false, initHeader) == 'true' then
             error('Argon is already connected!', 0)
         end
     end)
@@ -87,15 +90,42 @@ function httpHandler.connect(fail)
 
     if success then
         startSyncing(url)
+
+        if not title then
+            if game.Name:find('.rbxl') or game.Name:find('.rbxlx') then
+                response = ' - '..game.Name
+                title = response
+            else
+                success = pcall(function()
+                    response = ' - '..MarketPlaceService:GetProductInfo(game.PlaceId).Name
+                    title = response
+                end)
+
+                if not success then
+                    response = ''
+                end
+            end
+        else
+            response = title
+        end
+
+        if response ~= '' then
+            HttpService:PostAsync(url, response, Enum.HttpContentType.ApplicationJson, false, titleHeader)
+        end
     end
 
     return success, response
 end
 
 function httpHandler.disconnect()
+    local url = URL:format(Config.host, Config.port)
+    local header = {action = 'disconnect'}
+
     if thread then
         task.cancel(thread)
         thread = nil
+
+        HttpService:PostAsync(url, HttpService:JSONEncode(''), Enum.HttpContentType.ApplicationJson, false, header)
     end
 end